deposit: symlinks creation as celery task
