+++ cds/modules/deposit/api.py
@@ -41,7 +41,7 @@ from invenio_db import db
 from invenio_deposit.api import Deposit, has_status, preserve
 from invenio_files_rest.models import (Bucket, Location, MultipartObject,
                                        ObjectVersion, ObjectVersionTag,
-                                       as_bucket, FileInstance)
+                                       as_bucket)
 from invenio_jsonschemas import current_jsonschemas
 from invenio_pidstore.errors import (PIDInvalidAction, PIDDoesNotExistError,
                                      ResolverError)
@@ -56,6 +56,7 @@ from ..records.api import (CDSFileObject, CDSFilesIterator, CDSRecord)
 from ..records.minters import is_local_doi, report_number_minter
 from ..records.resolver import record_resolver
 from ..records.validators import PartialDraft4Validator
+from ..records.tasks import create_symlinks
 from ..webhooks.status import (ComputeGlobalStatus, get_deposit_events,
                                get_tasks_status_by_task,
                                iterate_events_results, merge_tasks_status)
@@ -693,11 +694,6 @@ class Video(CDSDeposit):
 
     _schema = 'deposits/records/videos/video/video-v1.0.0.json'
 
-    def __init__(self, *args, **kwargs):
-        """Init."""
-        super(Video, self).__init__(*args, **kwargs)
-        self._symlinks_creator = self.SymlinksCreator()
-
     @classmethod
     def create(cls, data, id_=None, **kwargs):
         """Create a video deposit.
@@ -791,6 +787,7 @@ class Video(CDSDeposit):
         if 'pid' in self['_deposit']:
             try:
                 _, previous_record = self.fetch_published()
+                previous_record = deepcopy(previous_record)
             except ResolverError:
                 # video not yet published
                 pass
@@ -800,15 +797,18 @@ class Video(CDSDeposit):
                                                      **kwargs)
         _, record_new = self.fetch_published()
 
-        # create file symlinks
-        self._symlinks_creator.create(previous_record, record_new)
-
         # update associated project
         video_published.project._update_videos(
             [video_build_url(video_old_id)],
             [record_build_url(record_new['recid'])]
         )
         video_published.project.commit()
+
+        # create file symlinks delayed (waiting the commit)
+        create_symlinks.s(
+            previous_record=previous_record, record_uuid=str(record_new.id)
+        ).apply_async(countdown=90)
+
         return video_published
 
     def edit(self, pid=None):
@@ -930,94 +930,6 @@ class Video(CDSDeposit):
             ObjectVersionTag.create_or_update(poster, 'media_type', 'image')
             # refresh object
             db.session.refresh(poster)
-
-    class SymlinksCreator(object):
-        """."""
-
-        def __init__(self):
-            """Init."""
-            self._symlinks_location = self._get_symlinks_location()
-
-        def create(self, prev_record, new_record):
-            """Create the video symlink with a human readable path."""
-            accepted_context_types = ['master']
-
-            # delete old symlink for previous record if any
-            if prev_record:
-                self._delete_prev_symlink(prev_record,
-                                          accepted_context_types)
-
-            for _file in new_record.get('_files', []):
-                if _file.get('context_type', '') in accepted_context_types:
-                    self._create_symlink_for_file(_file, new_record)
-
-        def _create_symlink_for_file(self, file, record):
-            # fetch file location
-            file_instance = FileInstance.get(file['file_id'])
-            # build the new symlink path
-            symlink_path = self._build_link_path(
-                self._symlinks_location, record, file['key'])
-            if file_instance and os.path.exists(file_instance.uri):
-                source = file_instance.uri
-
-                # build symlink path directories recursively
-                if not os.path.isdir(os.path.dirname(symlink_path)):
-                    os.makedirs(os.path.dirname(symlink_path))
-
-                info = "New symlink - source: {source}\nlink: {link}" \
-                    .format(source=source, link=symlink_path)
-                current_app.logger.info(info)
-
-                # delete the prev symlink for safety in case it points
-                # to another file or it is broken
-                self._delete_symlink(symlink_path)
-                # create the new symlink
-                os.symlink(source, symlink_path)
-
-        def _delete_prev_symlink(self, prev_record, accepted_context_types):
-            """Delete the symlink to the previous published file."""
-            for _file in prev_record.get('_files', []):
-                if _file.get('context_type', '') in accepted_context_types:
-                    link_name = self._build_link_path(self._symlinks_location,
-                                                      prev_record,
-                                                      _file['key'])
-                    self._delete_symlink(link_name)
-
-        @staticmethod
-        def _delete_symlink(link_name):
-            """Delete a symlink if it exists."""
-            if os.path.lexists(link_name):
-                info = "Deleting symlink\nlink: {link}\nWas pointing to: " \
-                       "{source}".format(link=link_name,
-                                         source=os.path.realpath(link_name))
-                current_app.logger.info(info)
-                os.unlink(link_name)
-
-        @staticmethod
-        def _build_link_path(symlinks_location, record, filename):
-            """Build the symlink path."""
-            report_number = record['report_number'][0] if \
-                len(record['report_number']) > 0 else record['report_number']
-            year = arrow.get(record['date']).year
-
-            return os.path.join(symlinks_location,
-                                record['type'],
-                                record['category'],
-                                str(year),
-                                report_number,
-                                filename)
-
-        @staticmethod
-        def _get_symlinks_location():
-            """Set the video symlinks location."""
-            links_dir_name = 'links'
-
-            # go one dir up to set the links location
-            # before: /path/to/videos/files
-            # after: /path/to/videos/links
-            videos_location = Location.get_by_name('videos')
-            base_no_last, _ = os.path.split(videos_location.uri)
-            return os.path.join(base_no_last, links_dir_name)
 
 
 project_resolver = Resolver(
